separate server module
